Move test to admin tests
